[#2129] Remove StandAloneReactiveTest
I'm not sure what's the purpose of this test, and why
it was working before.

But, it doesn't work anymore with the fix for #2129 because
some properties are missing and we are not setting the right
credentials.

I think it was one of first tests we inroduced when we have created
the project.